split the game.py file into multiple files
