+++ usolitaire/__init__.py
@@ -3,3 +3,8 @@
 __author__ = """Elias Dorneles"""
 __email__ = "eliasdorneles@gmail.com"
 __version__ = "1.1.0"
+
+from .game import Game
+from .card import Card
+from .deck import Deck
+from .exceptions import InvalidMove
