+++ usolitaire/exceptions.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+class InvalidMove(Exception):
+    """Raised to indicate an invalid move"""
